webkit
added the scripts for the webkit. the token for this expires ever hour. but we can go back in and check it manually as needed for now. i also was able to isolate the track uri . which i read we needed for play back . i currently have it connected to a random a tag i createdin the html. if its clicked it redirects to the spotify page. not exactly what we need but atleast now we have an on page way to access the music that weve been logging. i change the request limit to one for the time being just to have less data on the console log.